requirements.txt
annotated-types==0.6.0
anyio==4.3.0
boto3==1.34.48
botocore==1.34.48
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
jmespath==1.0.1
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.12.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
StrEnum==0.4.15
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==1.26.18
